requirements.txt
django_graphql_jwt==0.2.2
graphene_django==2.6.0
channels==2.3.0
graphene==2.1.8
Django==2.2.6
django-channels-graphql-ws==0.4.2
django-filter==2.2.0
django-cors-headers==3.1.1
graphql_relay==2.0.0